refactor tests for apps service
