+++ api-implementation/test/integ/apps/apps.developer.update.spec.ts
@@ -4,15 +4,14 @@ import path from 'path';
 import type { App, AppPatch } from "../../lib/generated/openapi";
 import { ApiError, AppStatus, AppsService } from "../../lib/generated/openapi";
 
-import * as setup from './apps.test.setup';
-import { AclProfile, Queue } from './apps.test.helpers';
+import * as setup from './common/test.setup';
+import { AclProfile, Queue } from './common/test.helpers';
 import {
   createRestDeliveryPointFromWebHook,
   verifyAclProfile,
   verifyMessageQueue,
   verifyRestDeliveryPoint,
-} from './apps.test.helpers';
-import { PlatformAPIClient } from '../../lib/api.helpers';
+} from './common/test.helpers';
 
 const scriptName: string = path.basename(__filename);
