+++ api-implementation/test/integ/integration.spec.ts
@@ -14,5 +14,5 @@ describe("apps", function() {
   require('./apps/apps.developer.create.spec.ts')
   require('./apps/apps.developer.delete.spec.ts')
   require('./apps/apps.developer.list.spec.ts')
-  require('./apps/apps.developer.patch.spec.ts')
+  require('./apps/apps.developer.update.spec.ts')
 });
